# Differ: write diff to the content store over bufio writer