@ktock 2021-10-04
> Could you post results with oci worker as well for comparison (just with the patch)?

Added the results with OCI worker.